requirements.txt
fastapi~=0.116.1
uvicorn
httpx~=0.28.1
numpy~=2.0.2
openai~=1.97.0
python-dotenv~=1.1.1